Turned values into int64.

+++ tx.go
@@ -92,7 +92,7 @@ func (api *API) GetTXConf(c appengine.Context, hash string) (conf TXConf, err er
 
 //TempNewTX creates a simple template transaction, suitable for
 //use in NewTX. Takes an input/output address and amount.
-func TempNewTX(c appengine.Context, inAddr string, outAddr string, amount int) (trans TX) {
+func TempNewTX(c appengine.Context, inAddr string, outAddr string, amount int64) (trans TX) {
 	trans.Inputs = make([]TXInput, 1)
 	trans.Outputs = make([]TXOutput, 1)
 	trans.Inputs[0].Addresses = make([]string, 1)
@@ -110,7 +110,7 @@ func TempNewTX(c appengine.Context, inAddr string, outAddr string, amount int) (
 //send from a multisig address (/series of public keys).
 //n represents the number of valid signatures required, and m
 //is derived from the number of pubkeys.
-func TempMultiTX(c appengine.Context, inAddr string, outAddr string, amount int, n int, pubkeys []string) (trans TX, err error) {
+func TempMultiTX(c appengine.Context, inAddr string, outAddr string, amount int64, n int, pubkeys []string) (trans TX, err error) {
 	m := len(pubkeys)
 	if inAddr != "" && outAddr != "" {
 		err = errors.New("TempMultiTX: Can't have both inAddr and outAddr != \"\"")
